+++ R/writemeteorologygrid.R
@@ -10,7 +10,41 @@ writemeteorologygrid<-function(object, file, dates = NULL, format = "netCDF", ad
                               nc=nc)
   .closeNetCDF(file,nc)
 }
-
+writemeteorologygridpixel<-function(file, index, data) {
+  nc = .openaddNetCDF(file=file, verbose=FALSE)
+  gt = .readgridtopologyNetCDF(nc)
+  gdates = .readdatesNetCDF(nc)
+  ny = gt@cells.dim[2]
+  nt = length(gdates)
+  cv = coordinatevalues(gt)
+  cci = coordinates(gt)[index,]
+  i = which(cv[[1]]==cci[1])
+  j = which(cv[[2]]==cci[2])
+  varMeanTemp = nc$var$MeanTemperature
+  varMinTemp = nc$var$MinTemperature
+  varMaxTemp = nc$var$MaxTemperature
+  varPrec = nc$var$Precipitation
+  varMeanRH = nc$var$MeanRelativeHumidity
+  varMinRH = nc$var$MinRelativeHumidity
+  varMaxRH = nc$var$MaxRelativeHumidity
+  varRad = nc$var$Radiation
+  varWindSpeed = nc$var$WindSpeed
+  varWindDirection = nc$var$WindDirection
+  varPET = nc$var$PET
+  if("MeanTemperature" %in% names(data)) .putvardatapixel(nc,ny, nt, varMeanTemp, i, j, data$MeanTemperature)
+  if("MinTemperature" %in% names(data)) .putvardatapixel(nc,ny, nt, varMinTemp, i, j, data$MinTemperature)
+  if("MaxTemperature" %in% names(data)) .putvardatapixel( nc, ny, nt, varMaxTemp, i, j, data$MaxTemperature)
+  if("Precipitation" %in% names(data)) .putvardatapixel( nc, ny, nt, varPrec, i, j, data$Precipitation)
+  if("MeanRelativeHumidity" %in% names(data)) .putvardatapixel( nc, ny, nt, varMeanRH, i, j, data$MeanRelativeHumidity)
+  if("MinRelativeHumidity" %in% names(data)) .putvardatapixel( nc, ny, nt, varMinRH, i, j, data$MinRelativeHumidity)
+  if("MaxRelativeHumidity" %in% names(data)) .putvardatapixel( nc, ny, nt, varMaxRH, i, j, data$MaxRelativeHumidity)
+  if("Radiation" %in% names(data)) .putvardatapixel( nc, ny, nt, varRad, i, j, data$Radiation)
+  if("WindSpeed" %in% names(data)) .putvardatapixel( nc, ny, nt, varWindSpeed, i, j, data$WindSpeed)
+  if("WindDirection" %in% names(data)) .putvardatapixel( nc, ny, nt, varWindDirection, i, j, data$WindDirection)
+  if("PET" %in% names(data)) .putvardatapixel( nc, ny, nt, varPET, i, j, data$PET)
+  
+  .closeNetCDF(grid, nc, verbose=FALSE)
+}
 writemeteorologypixels<-function(object, file, dates = NULL, format = "netCDF", add=FALSE, overwrite = FALSE) {
   if(!inherits(object,"SpatialPixelsMeteorology")) stop("'object' has to be of class 'SpatialPixelsMeteorology'.")
   if(is.null(dates)) dates = object@dates