+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 0.8.7
 - New function 'readmeteorologygridpoints'
+- New function 'writemeteorologygridpixel'
 - New functions 'mergegrids' and 'mergepoints'
 - Reading multiple grid/pixels files